backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.7.1
python-multipart==0.0.6
pydantic==2.7.4
psycopg2-binary==2.9.9
sentence-transformers
huggingface-hub
pgvector==0.2.4
sqlalchemy==2.0.23
langchain==0.2.17
langchain-community==0.2.19
langchain-google-genai
unstructured
pypdf